requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
exceptiongroup==1.2.2
fastapi==0.112.1
frozenlist==1.4.1
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
multidict==6.0.5
priority==2.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
strawberry-graphql==0.237.3
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.2
wsproto==1.2.0
yarl==1.9.4
